+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.5.0] - 2025-05-20
+### Added
+- Prompt to add default outflow categories when none exist.
+### Changed
+- Inflow categories no longer appear in transaction or category lists and are ignored in budgets.
 ## [0.4.0] - 2025-05-19
 ### Changed
 - New transaction screen elements animate sequentially when opening.